Add files via upload
This is a Project to Monitor the Battery Status on devices.

Recently, the battery life has been shown to be exposed to too much power and over-charging. This project sets up an alarm system to notify user once the battery is fully charged.

There's more work to be done though just some ideas for now.